Programa documentado y con readme creado para la explicación inicial

+++ Planets/src/mainclass/MainClass.java
@@ -1,6 +1,6 @@
 package mainclass;
 
-import elements.Element;
+import elements.SpaceElement;
 import elements.Planet;
 import elements.Satellite;
 import filemanipulation.HtmlFileControl;
@@ -15,11 +15,9 @@ import user.User;
 import visualfront.ConsoleColors;
 import visualfront.Menu;
 import visualfront.Paint;
-
 /**
  * @author Alex Guirao López <aguiraol2021@cepnet.net>
  */
-//Alt + shift + m => Comando para encapsular metodos.
 public class MainClass 
 {
     final static String [] MAIN_MENU_LINES = new String [] {" --- Menú principal ---",
@@ -33,19 +31,24 @@ public class MainClass
                                                                                         ConsoleColors.RED+"-1 => Eliminar (DANGER!)"};
     
     /**
-     * Bienvenido al programa.
-     * Sigue los (números) con el orden de la lógica del programa.
+     * Programa de gestión para planetas y satélites.
+     * Permite registrar y leer información sobre ellos.
+     * También permite generar y ejecutar archivos html 
+     * con los datos correspondientes para ver 
+     * la información más detallada.
      * @param args 
      */
     public static void main(String[] args) 
     {
-        Sound.startAudio();
-        Paint.drawMainImage(); //(0.) Dibuja la imagen de presentación.
+        Sound.startAudio(); 
+        Paint.drawMainImage(); 
+        Paint.breakLine();
         Menu mainMenu = new Menu(MAIN_MENU_LINES);
+        
         int userAns;
         
         do{
-            //(1.) Se muestra el menú principal para que el usuario elija una opción.
+            //Se muestra el menú principal para que el usuario elija una opción.
             mainMenu.showMenu();
             userAns=Keyboard.readInt();
             switch(userAns)
@@ -80,16 +83,18 @@ public class MainClass
                 case -1:eraseAllInfo(); //ELIMINA TODA LA INFORMACIÓN
                     break;
                     
-                default: System.out.println(ConsoleColors.RED+"Opción errónea");
+                case 0: //Salir
+                    System.out.println("¡No olvides que tus datos han sido registrados!");
+                    System.out.println("Vuelve a acceder cuando te apetezca ;)");
+                    break;
+                    
+                default: 
+                    System.out.println(ConsoleColors.RED+"Opción errónea");
+                    break;
             }
-            
-            
-            
             Paint.breakLine();
             
         }while(userAns!=0);
-        System.out.println("¡No olvides que tus datos han sido registrados!");
-        System.out.println("Vuelve a acceder cuando te apetezca ;)");
     }
 
     //========================MÉTODOS PRINCIPALES=========================
@@ -101,8 +106,8 @@ public class MainClass
     private static void insertPlanet() 
     {
         System.out.println(ConsoleColors.CYAN+"    --Introducir planeta--");
-        Planet planet = User.createPlanet(); //(2.) El usuario crea un planeta.
-        PlanetFileControl.writePlanet(planet);//(3.) La info del planeta se registra en el fichero.
+        Planet planet = User.createPlanet(); //El usuario crea un planeta.
+        PlanetFileControl.writePlanet(planet);//La info del planeta se registra en el fichero.
     }
     
     /**
@@ -114,10 +119,26 @@ public class MainClass
         if (PlanetFileControl.getPlanetAmount()>0)
         {
             System.out.println(ConsoleColors.PURPLE+"   --Introducir satélite--");
-            Satellite satellite = User.createSatellite();//(4.) El usuario crea un satélite.
-            SatelliteFileControl.writeSatellite(satellite); //(5.) La info del satélite se registra en el fichero.
+            Satellite satellite = User.createSatellite(); //El usuario crea un satélite.
+            SatelliteFileControl.writeSatellite(satellite); //La info del satélite se registra en el fichero.
         }else{
             System.out.println(ConsoleColors.RED+"Inserta mínimo un planeta antes de satélites");
+        }
+    }
+    
+     /**
+     * Muestra una lista numerada con nombres de un elemento (satélite o planeta)
+     * en el orden de su archivo binario correspondiente.
+     * @param elementList lista de elementos a mostrar.
+     */
+    public static void showElementNames(ArrayList<SpaceElement> elementList)
+    {
+        int nameIndex=1;
+
+        for (SpaceElement element:elementList)
+        {
+            System.out.println(nameIndex+"- "+element.getName());
+            nameIndex++;
         }
     }
     
@@ -127,14 +148,17 @@ public class MainClass
      */
     private static void showPlanet() 
     {
-        ArrayList planetList = PlanetFileControl.readPlanetList();
-        showElementNames(planetList);
+        //Guarda en una lista los planetas del archivo en orden.
+        ArrayList planetList = PlanetFileControl.readPlanetList(); 
+        showElementNames(planetList); //Muestra los nombres de los planetas al usuario.
         
         if (PlanetFileControl.getPlanetAmount()>0)
         {
+            /*El usuario selecciona el planeta escribiendo el número de su posición 
+            (-1 para igualar al sistema de posicionamiento de los registros).*/
             Planet planet=PlanetFileControl.readPlanet(User.selectPlanet()-1);
-            String planetInfo= planet.toString();//(6.) La info del planeta se recoge.
-            System.out.println(planetInfo);//(7.) Muestra la info del planeta.
+            String planetInfo= planet.toString();//La info del planeta se recoge.
+            System.out.println(planetInfo);//Muestra la info del planeta.
         }else{
             System.out.println(ConsoleColors.RED+"No hay planetas todavía");
         }
@@ -146,14 +170,17 @@ public class MainClass
      */
     private static void showSatellite() 
     {
+        //Guarda en una lista los satélites del archivo en orden.
         ArrayList satelliteList = SatelliteFileControl.readSatelliteList();
         showElementNames(satelliteList);
         
         if (SatelliteFileControl.getSatelliteAmount()>0)
         {
+            /*El usuario selecciona el satélite escribiendo el número de su posición 
+            (-1 para igualar al sistema de posicionamiento de los registros).*/
             Satellite planet=SatelliteFileControl.readSatellite(User.selectSatellite()-1);
-            String satelliteInfo= planet.toString();//(6.) La info del planeta se recoge.
-            System.out.println(satelliteInfo);//(7.) Muestra la info del planeta.
+            String satelliteInfo= planet.toString();//La info del planeta se recoge.
+            System.out.println(satelliteInfo);//Muestra la info del planeta.
         }else{
             System.out.println(ConsoleColors.RED+"No hay satélites todavía");
         }
@@ -165,11 +192,13 @@ public class MainClass
      */
      private static void generateWeb()
     {
+        //Guarda en una lista los planetas del archivo en orden.
         ArrayList planetList = PlanetFileControl.readPlanetList();
         showElementNames(planetList);
+        
         if (PlanetFileControl.getPlanetAmount()>0)
         {
-            //(10.) Genera archivo HTML del planeta seleccionado.
+            //Genera archivo HTML del planeta seleccionado.
             HtmlFileControl.generateHTMLFile(User.selectPlanet()-1);
             System.out.println(ConsoleColors.GREEN+"Archivo HTML generado correctamente.");
         }else{
@@ -178,7 +207,7 @@ public class MainClass
     }
      
      /**
-      * Ejecuta una web.
+      * Ejecuta un archivo HTML para abrir la web del planeta en el navegador.
       * (Opción 6 del programa)
       */
      private static void executeWeb()
@@ -190,12 +219,13 @@ public class MainClass
             int loopIndex=1;
             int userAns;
 
-            do{
+            do{ //Muestra los archivos html existentes.
                 for (File f:file)
                {
                     System.out.println(loopIndex+"- "+f.getName());
                     loopIndex++;
                 }
+                
                userAns=User.insertUnsignedInt("Selecciona un archivo");
 
                if (userAns>file.length)
@@ -204,12 +234,11 @@ public class MainClass
                }
             }while(userAns>file.length);
 
-            HtmlFileControl.executeFile(userAns);
+            HtmlFileControl.executeFile(userAns); //Ejecuta el archivo HTML.
             System.out.println(ConsoleColors.GREEN+"Se ha ejecutado el archivo correctamente.");
          }else{
              System.out.println(ConsoleColors.RED+"Todavía no existen archivos HTML...");
          }
-         
      }
      
     /**
@@ -218,30 +247,25 @@ public class MainClass
      */
     public static void eraseAllInfo()
     {
-        PlanetFileControl.file.delete();
-        SatelliteFileControl.file.delete();
+        char userAns;
+        //Avisos de seguridad antes de eliminar los datos.
+        System.out.println(ConsoleColors.RED+"¿Estás seguro de eliminar todos los datos?");
+        System.out.println(ConsoleColors.RED+"Introduce S para aceptar. Cualquier otro carácter para cancelar. ");
+        userAns=Keyboard.readChar();
         
-        File [] file = HtmlFileControl.file.listFiles();
-        
-        for (File f:file)
+        if (userAns==Character.toLowerCase('s'))
         {
-            f.delete();
-        }
-    }
-    
-       /**
-     * Muestra una lista numerada con nombres de algún elemento
-     * en el orden de su archivo binario correspondiente.
-     * @param elementList lista de elementos a mostrar.
-     */
-    public static void showElementNames(ArrayList<Element> elementList)
-    {
-        int nameIndex=1;
+            //Elimina todos los ficheros de datos.
+            PlanetFileControl.file.delete();
+            SatelliteFileControl.file.delete();
+            File [] file = HtmlFileControl.file.listFiles();
 
-        for (Element element:elementList)
-        {
-            System.out.println(nameIndex+"- "+element.getName());
-            nameIndex++;
+            for (File f:file)
+            {
+                f.delete();
+            }
+            
+            System.out.println(ConsoleColors.RED+"Los datos han sido eliminados correctamente.");
         }
     }
 }
